+++ static_report/src/components/shared/Charts/index.ts
@@ -3,3 +3,9 @@ export * from './CategoricalBarChart';
 export * from './FlatBoxPlotChart';
 export * from './FlatStackedBarChart';
 export * from './HistogramChart';
+export {
+  type ChartKind,
+  getBoxPlotKeyData,
+  getChartKindByColumnType,
+  getChartUnavailMsg,
+} from './utils';
