+++ static_report/src/components/shared/Widgets/ChartTabsWidget.tsx
@@ -11,8 +11,11 @@ import {
 } from '@chakra-ui/react';
 import isNumber from 'lodash/isNumber';
 import { ColumnSchema } from '../../../sdlc/single-report-schema';
-import { getDataChart, renderChartUnavailableMsg } from '../Charts/utils';
-import { ChartKind } from '../../../utils/transformers';
+import {
+  ChartKind,
+  getDataChart,
+  renderChartUnavailableMsg,
+} from '../Charts/utils';
 import { ChartContainer } from '../Charts/ChartContainer';
 import { TEXTLENGTH } from '../Columns/constants';
