+++ static_report/src/components/shared/index.ts
@@ -1,7 +1,7 @@
 // export * from './Assertions';
-export * from './Charts';
-// export * from './Columns';
 // export * from './Layouts';
+export * from './Charts';
+export * from './Columns';
 // export * from './Modals';
 // export * from './Tables';
 // export * from './Widgets';
